// src/app/components/delivery-dashboard/model/DeliveryOrders.model.ts
export interface DeliveryOrdersModel {
    orderId:     number;
    id:          number;
    addressName: string;
    userId:      number;
    phoneNumber: string;
    city:        string;
    zipCode:     number;
    details:     string;
    street:      string;
    lat:         number;
    lng:         number;
    firstName:   string;
    lastName:    string;
}